IGS-19 refactor: record 구조 단순화
