+++ src/main/java/com/inglo/giggle/security/application/dto/request/SignUpDefaultOwnerRequestDto.java
@@ -1,8 +1,8 @@
 package com.inglo.giggle.security.application.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.inglo.giggle.address.dto.AddressDto;
-import com.inglo.giggle.security.application.dto.inner.SignUpDefaultOwnerOwnerInfo;
+import com.inglo.giggle.address.dto.request.AddressRequestDto;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record SignUpDefaultOwnerRequestDto(
@@ -14,7 +14,7 @@ public record SignUpDefaultOwnerRequestDto(
         SignUpDefaultOwnerOwnerInfo signUpDefaultOwnerOwnerInfo,
 
         @JsonProperty("address")
-        AddressDto address,
+        AddressRequestDto address,
 
         @JsonProperty("marketing_allowed")
         @NotNull(message = "마케팅 정보 활용 동의 여부를 입력해주세요.")
@@ -24,4 +24,21 @@ public record SignUpDefaultOwnerRequestDto(
         @NotNull(message = "알림 수신 동의 여부를 입력해주세요.")
         Boolean notificationAllowed
 ) {
+        public record SignUpDefaultOwnerOwnerInfo(
+                @JsonProperty("company_name")
+                @NotBlank(message = "회사/점포명을 입력해주세요.")
+                String companyName,
+
+                @JsonProperty("owner_name")
+                @NotBlank(message = "이름을 입력해주세요.")
+                String ownerName,
+
+                @JsonProperty("company_registration_number")
+                @NotBlank(message = "사업자 등록번호를 입력해주세요.")
+                String companyRegistrationNumber,
+
+                @JsonProperty("phone_number")
+                @NotBlank(message = "전화번호를 입력해주세요.")
+                String phoneNumber
+        ){}
 }
