+++ src/main/java/com/inglo/giggle/security/application/dto/request/SignUpDefaultUserRequestDto.java
@@ -1,9 +1,11 @@
 package com.inglo.giggle.security.application.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.inglo.giggle.address.dto.AddressDto;
-import com.inglo.giggle.security.application.dto.inner.SignUpDefaultUserUserInfo;
+import com.inglo.giggle.address.dto.request.AddressRequestDto;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+
+import java.time.LocalDate;
 
 public record SignUpDefaultUserRequestDto(
         @JsonProperty("temporary_token")
@@ -14,7 +16,7 @@ public record SignUpDefaultUserRequestDto(
         SignUpDefaultUserUserInfo signUpDefaultUserUserInfo,
 
         @JsonProperty("address")
-        AddressDto address,
+        AddressRequestDto address,
 
         @JsonProperty("marketing_allowed")
         @NotNull(message = "마케팅 정보 활용 동의 여부를 입력해주세요.")
@@ -28,4 +30,33 @@ public record SignUpDefaultUserRequestDto(
         @NotNull(message = "사용 언어 선택은 필수입니다.")
         String language
 ) {
+        public record SignUpDefaultUserUserInfo(
+                @JsonProperty("first_name")
+                @NotBlank(message = "first name을 입력해주세요.")
+                String firstName,
+
+                @JsonProperty("last_name")
+                @NotBlank(message = "last name을 입력해주세요.")
+                String lastName,
+
+                @JsonProperty("gender")
+                @NotNull(message = "성별을 선택해주세요.")
+                String gender,
+
+                @JsonProperty("birth")
+                @NotNull(message = "생년월일을 입력해주세요.")
+                LocalDate birth,
+
+                @JsonProperty("nationality")
+                @NotBlank(message = "국적을 입력해주세요")
+                String nationality,
+
+                @JsonProperty("visa")
+                @NotNull(message = "비자를 입력해주세요.")
+                String visa,
+
+                @JsonProperty("phone_number")
+                @NotBlank(message = "전화번호를 입력해주세요.")
+                String phoneNumber
+        ) {}
 }
